Revert "Move citool.go to citool/main.go"
This reverts commit 31b675556412496f9eed93b29015dfc1624b77e2.
