+++ .gitignore
@@ -15,4 +15,3 @@
 
 # Default directory for temporary Circle CI data
 circleci_data
-bin/
