+++ src/citool/analyze.go
