+++ src/citool/filter.go
