+++ src/citool/jobstatus.go
